Rename clientapp & clientapp13.
Rename clientapp to clientapp-prior.
Rename clientapp13 to clientapp.
In commit 4b1b490d9a, we started building "clientapp13",
which would be the upgrade of clientapp from Angular 12 to 13.
It was also the opportunity to:
 * do a lot of refactoring of the clientapp code
 * get rid of any unused code
 * ensure that only needed NPM packages are referenced
   and that they are the latest compatable versions
We built clientapp13 step-by-step, moving over small self-contained
parts from clientapp. The git history of these changes also
documents a process that we can use in the future when large-scale
dependencies need to be upgraded safely.
Doing this via "ng upgrade ..." is not always the safest.